backend/requirements.txt
# Core Django and DRF
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
setuptools==69.5.1

# Database
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# File handling
Pillow==10.2.0

# Production server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0

# CORS
django-cors-headers==4.3.1

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
django-extensions==3.2.3
